fix(unusedRoutes): remove unused routes and controllers
