--- api/controllers/JTopoCaveController.js
@@ -1,8 +1,0 @@
-/**
- * JTopoCave.js
- *
- * @description :: jTopoCave controller imported from localhost MySql server at 15/11/2016 19:18:1.
- * @docs        :: http://sailsjs.org/#!documentation/controllers
- */
-
-module.exports = {};
