--- api/controllers/JTopoEntryController.js
@@ -1,8 +1,0 @@
-/**
- * JTopoEntry.js
- *
- * @description :: jTopoEntry controller imported from localhost MySql server at 15/11/2016 19:18:6.
- * @docs        :: http://sailsjs.org/#!documentation/controllers
- */
-
-module.exports = {};
