+++ api/controllers/SearchController.js
@@ -1,59 +1,131 @@
 /**
  */
 
+const ElasticSearch = require('../services/ElasticsearchService');
+
+const advancedSearchTypes = ['massifs', 'entrances', 'grottos', 'documents'];
+
 module.exports = {
-  findAll: (req, res, converter) => {
-    const apiControl = req.options.api;
-    const parameters = {};
-    let { limit } = apiControl;
-    let skip = 0;
-
-    if (req.param('name')) {
-      parameters.name = {
-        like: `%${req.param('name')}%`,
-      };
+  /**
+   * Perform a quick search using multiple URL parameters :
+   * - resourceType:  string, entity type to search for (@see advancedSearchTypes) (FACULTATIVE).
+   *                  By default, search on all entities
+   **/
+  search: (req, res) => {
+    // Store every params in the url and check if the query parameter exists
+    const paramsURl = req.query;
+    if (!paramsURl.query) {
+      return res.badRequest();
     }
 
-    const maxRange = limit;
-    const range = req.param('range');
-
-    if (range !== undefined) {
-      const splitRange = range.split('-');
-      limit = splitRange[1] - splitRange[0];
-      skip = splitRange[0];
+    // By default, the query asked will send every information.
+    // We can limit these information just by adding a "complete" parameter to false in the query.
+    let complete = true;
+    if (paramsURl.complete === 'false' || paramsURl.complete === false) {
+      complete = false;
     }
 
-    // TODO : to adapt when authentication will be implemented
-    parameters.isPublic = 'YES';
+    const params = {
+      searchedItem: `Search entity with the following query ${req.query.query}`,
+    };
 
-    TEntrance.count(parameters).exec((error, total) => {
-      TEntrance.find(parameters)
-        .sort('id ASC')
-        .limit(limit)
-        .skip(skip)
-        .exec((err, foundEntry) => {
-          const params = {
-            controllerMethod: 'SearchController.findAll',
-            notFoundMessage: 'No entries found.',
-            searchedItem: apiControl.entity,
-            total,
-            url: req.originalUrl,
-            maxRange,
-            limit,
-            skip,
-          };
-
-          // only search for entries at this time
-          // return ControllerService.treat(req, err, foundCave.concat(foundEntry).concat(foundGrotto), params, res);
+    // Use the Elasticsearch Service to do the search according to the parameters of the URL
+    return ElasticSearch.searchQuery(paramsURl)
+      .then((results) => {
+        if (complete) {
           return ControllerService.treatAndConvert(
             req,
-            err,
-            foundEntry,
+            undefined,
+            results,
             params,
             res,
-            converter,
+            MappingV1Service.convertToCompleteSearchResult,
           );
-        });
-    });
+        }
+        return ControllerService.treatAndConvert(
+          req,
+          undefined,
+          results,
+          params,
+          res,
+          MappingV1Service.convertEsToSearchResult,
+        );
+      })
+      .catch((err) => {
+        return ControllerService.treatAndConvert(
+          req,
+          err,
+          undefined,
+          params,
+          res,
+          MappingV1Service.convertEsToSearchResult,
+        );
+      });
+  },
+
+  /**
+   * Perform an advanced search using multiple URL parameters :
+   * - resourceType: string, entity type to search for (@see advancedSearchTypes) (NEEDED)
+   * - complete: bool, determine if the results must be returned in their entirely or just their id and name (default = false) (FACULTATIVE)
+   * - matchAllFields: bool, determine if the results need to match all the fields (logic AND) or at least one of them (logic OR) (default = true) (FACULTATIVE)
+   */
+  advancedSearch: (req, res) => {
+    // Store every params in the url and check if there is the type parameter
+    const paramsURL = req.query;
+
+    if (!advancedSearchTypes.includes(paramsURL.resourceType)) {
+      return res.badRequest();
+    }
+
+    // By default, the query asked will send every information.
+    let complete = true;
+    if (paramsURL.complete && paramsURL.complete === 'false') {
+      complete = false;
+    }
+
+    // Cast matchAllFields to bool
+    if (paramsURL.matchAllFields && paramsURL.matchAllFields === 'false') {
+      paramsURL.matchAllFields = false;
+    } else if (
+      paramsURL.matchAllFields &&
+      paramsURL.matchAllFields === 'true'
+    ) {
+      paramsURL.matchAllFields = true;
+    }
+
+    const params = {};
+
+    // Use the Elasticsearch Service to do the search according to the parameters of the URL
+    return ElasticSearch.advancedSearchQuery(paramsURL)
+      .then((results) => {
+        if (complete) {
+          return ControllerService.treatAndConvert(
+            req,
+            undefined,
+            results,
+            params,
+            res,
+            MappingV1Service.convertToCompleteSearchResult,
+          );
+        }
+        return ControllerService.treatAndConvert(
+          req,
+          undefined,
+          results,
+          params,
+          res,
+          MappingV1Service.convertEsToSearchResult,
+        );
+      })
+      .catch((err) => {
+        return ControllerService.treatAndConvert(
+          req,
+          err,
+          undefined,
+          params,
+          res,
+          MappingV1Service.convertEsToSearchResult,
+        );
+      });
   },
 };
