+++ logstash.conf
@@ -45,7 +45,7 @@ input {
       WHERE is_validated = true
       GROUP BY doc.id, t_description.id
     "
-    type => "doc"
+    type => "document"
   }
   jdbc { 
     jdbc_connection_string => "${JDBC_POSTGRESQL}"
@@ -169,7 +169,7 @@ output {
     }
   }
 
-  if [type] == "doc" {
+  if [type] == "document" {
     elasticsearch {
       hosts => "${ES_HOSTS}"
       index => "doc-index"
